// patchnotes-web/src/api/generated/model/packageHealthDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * PatchNotes.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PackageHealthDto {
  id: string;
  name: string;
  githubOwner: string;
  githubRepo: string;
  consecutiveFailures?: number;
  /** @nullable */
  lastFailureAt?: string | null;
  /** @nullable */
  lastFailureMessage?: string | null;
  isSyncDisabled?: boolean;
  /** @nullable */
  lastFetchedAt?: string | null;
}